Move AuthTenant and co to utils
