+++ lib/Utils/AuthTenant.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ESN\DAV\Auth\Backend;
+namespace ESN\Utils;
 
 class AuthTenant
 {
